Revert "move tests"
This reverts commit 6247d3201701cdf3eb8de0d4ae5fd84c826c2d47.
